CLN: move fire and segment parameters from PyFibreRunner
