+++ pyfibre/shg_pl_trans/shg_analyser.py
@@ -5,7 +5,7 @@ import logging
 import pandas as pd
 from skimage.exposure import equalize_adapthist
 
-from traits.api import Instance, List, Any, Tuple
+from traits.api import Instance, List, Any, Tuple, Dict
 
 from pyfibre.model.core.base_multi_image_analyser import (
     BaseMultiImageAnalyser)
@@ -35,6 +35,12 @@ class SHGAnalyser(BaseMultiImageAnalyser):
     #: Reference to multi image under analysis
     multi_image = Instance(SHGImage)
 
+    #: Parameters used for FIRE algorithm
+    fire_parameters = Dict()
+
+    #: Parameters used for segmentation
+    segment_parameters = Dict()
+
     #: Reference to networkx Graph generated by analysis
     _network = Any
 
@@ -49,6 +55,23 @@ class SHGAnalyser(BaseMultiImageAnalyser):
 
     #: Reference to metric DataFrames generated by analysis
     _databases = Tuple()
+
+    def _fire_parameters_default(self):
+        return {
+            'nuc_thresh': 2,
+            'nuc_radius': 11,
+            'lmp_thresh': 0.15,
+            'angle_thresh': 70,
+            'r_thresh': 7
+        }
+
+    def _segment_parameters_default(self):
+        return {
+            'min_fibre_size': 100,
+            'min_fibre_frac': 0.1,
+            'min_cell_size': 200,
+            'min_cell_frac': 0.01,
+        }
 
     @property
     def data_path(self):
@@ -192,7 +215,7 @@ class SHGAnalyser(BaseMultiImageAnalyser):
 
     @log_time(message='NETWORK EXTRACTION')
     def network_analysis(
-            self, sigma, alpha, scale, p_denoise, fire_parameters):
+            self, sigma, alpha, scale, p_denoise):
         """Perform FIRE algorithm on image and save networkx
         objects for further analysis
 
@@ -229,12 +252,12 @@ class SHGAnalyser(BaseMultiImageAnalyser):
             scale=scale,
             sigma=sigma,
             alpha=alpha,
-            **fire_parameters)
+            **self.fire_parameters)
 
         self._fibre_networks = fibre_network_assignment(self._network)
 
     @log_time(message='SEGMENTATION')
-    def segmentation_analysis(self, scale, segment_parameters):
+    def segmentation_analysis(self, scale):
         """Segment image into regions
 
         Parameters
@@ -248,7 +271,7 @@ class SHGAnalyser(BaseMultiImageAnalyser):
             self.multi_image.segmentation_algorithm(
                 self._fibre_networks,
                 scale=scale,
-                **segment_parameters
+                **self.segment_parameters
             )
         )
